+++ modules/passkey/test/SafeWebAuthnSigner.spec.ts
@@ -2,7 +2,8 @@ import { expect } from 'chai'
 import { deployments, ethers } from 'hardhat'
 
 import * as ERC1271 from './utils/erc1271'
-import { DUMMY_AUTHENTICATOR_DATA, base64UrlEncode, encodeWebAuthnSigningMessage, getSignatureBytes } from './utils/webauthn'
+import { DUMMY_AUTHENTICATOR_DATA, base64UrlEncode, getSignatureBytes } from '../src/utils/webauthn'
+import { encodeWebAuthnSigningMessage } from './utils/webauthnShim'
 
 describe('SafeWebAuthnSigner', () => {
   const setupTests = deployments.createFixture(async () => {
@@ -39,8 +40,8 @@ describe('SafeWebAuthnSigner', () => {
         origin: 'https://safe.global',
       }
 
-      const r = ethers.id('signature.r')
-      const s = ethers.id('signature.s')
+      const r = BigInt(ethers.id('signature.r'))
+      const s = BigInt(ethers.id('signature.s'))
 
       const signature = getSignatureBytes({
         authenticatorData: DUMMY_AUTHENTICATOR_DATA,
@@ -73,8 +74,8 @@ describe('SafeWebAuthnSigner', () => {
         origin: 'https://safe.global',
       }
 
-      const r = ethers.id('signature.r')
-      const s = ethers.id('signature.s')
+      const r = BigInt(ethers.id('signature.r'))
+      const s = BigInt(ethers.id('signature.s'))
 
       const signature = getSignatureBytes({
         authenticatorData: DUMMY_AUTHENTICATOR_DATA,
@@ -102,17 +103,19 @@ describe('SafeWebAuthnSigner', () => {
       const data = ethers.toUtf8Bytes('some data to sign')
       const dataHash = ethers.keccak256(data)
 
-      const authenticatorData = ethers.solidityPacked(
-        ['bytes32', 'uint8', 'uint32'],
-        [
-          ethers.toBeHex(ethers.MaxUint256),
-          0, // no flags
-          0xffffffff, // signCount
-        ],
+      const authenticatorData = ethers.getBytes(
+        ethers.solidityPacked(
+          ['bytes32', 'uint8', 'uint32'],
+          [
+            ethers.toBeHex(ethers.MaxUint256),
+            0, // no flags
+            0xffffffff, // signCount
+          ],
+        ),
       )
 
-      const r = ethers.id('signature.r')
-      const s = ethers.id('signature.s')
+      const r = BigInt(ethers.id('signature.r'))
+      const s = BigInt(ethers.id('signature.s'))
 
       const signature = getSignatureBytes({
         authenticatorData,
