+++ modules/passkey/test/SafeWebAuthnSignerFactory.spec.ts
@@ -2,7 +2,8 @@ import { expect } from 'chai'
 import { deployments, ethers } from 'hardhat'
 
 import * as ERC1271 from './utils/erc1271'
-import { DUMMY_AUTHENTICATOR_DATA, base64UrlEncode, encodeWebAuthnSigningMessage, getSignatureBytes } from './utils/webauthn'
+import { DUMMY_AUTHENTICATOR_DATA, base64UrlEncode, getSignatureBytes } from '../src/utils/webauthn'
+import { encodeWebAuthnSigningMessage } from './utils/webauthnShim'
 
 describe('SafeWebAuthnSignerFactory', () => {
   const setupTests = deployments.createFixture(async () => {
@@ -95,8 +96,8 @@ describe('SafeWebAuthnSignerFactory', () => {
         origin: 'https://safe.global',
       }
 
-      const r = ethers.id('signature.r')
-      const s = ethers.id('signature.s')
+      const r = BigInt(ethers.id('signature.r'))
+      const s = BigInt(ethers.id('signature.s'))
       const x = ethers.id('publicKey.x')
       const y = ethers.id('publicKey.y')
 
@@ -128,8 +129,8 @@ describe('SafeWebAuthnSignerFactory', () => {
         origin: 'https://safe.global',
       }
 
-      const r = ethers.id('signature.r')
-      const s = ethers.id('signature.s')
+      const r = BigInt(ethers.id('signature.r'))
+      const s = BigInt(ethers.id('signature.s'))
       const x = ethers.id('publicKey.x')
       const y = ethers.id('publicKey.y')
 
@@ -156,17 +157,19 @@ describe('SafeWebAuthnSignerFactory', () => {
       const { factory, mockVerifier } = await setupTests()
 
       const dataHash = ethers.id('some data to sign')
-      const authenticatorData = ethers.solidityPacked(
-        ['bytes32', 'uint8', 'uint32'],
-        [
-          ethers.toBeHex(ethers.MaxUint256),
-          0, // no flags
-          0xffffffff, // signCount
-        ],
+      const authenticatorData = ethers.getBytes(
+        ethers.solidityPacked(
+          ['bytes32', 'uint8', 'uint32'],
+          [
+            ethers.toBeHex(ethers.MaxUint256),
+            0, // no flags
+            0xffffffff, // signCount
+          ],
+        ),
       )
 
-      const r = ethers.id('signature.r')
-      const s = ethers.id('signature.s')
+      const r = BigInt(ethers.id('signature.r'))
+      const s = BigInt(ethers.id('signature.s'))
       const x = ethers.id('publicKey.x')
       const y = ethers.id('publicKey.y')
