+++ modules/passkey/test/userstories/OffchainPasskeyVerification.spec.ts
@@ -1,9 +1,10 @@
-import { buildSignatureBytes } from '@safe-global/safe-4337/src/utils/execution'
 import { expect } from 'chai'
 import { deployments, ethers } from 'hardhat'
 
 import * as ERC1271 from '../utils/erc1271'
-import { WebAuthnCredentials, decodePublicKey, encodeWebAuthnSignature } from '../utils/webauthn'
+import { WebAuthnCredentials } from '../../test/utils/webauthnShim'
+import { decodePublicKey, encodeWebAuthnSignature } from '../../src/utils/webauthn'
+import { buildSignatureBytes } from '@safe-global/safe-4337/src/utils/execution'
 
 /**
  * User story: Off-chain Passkey Signature Verification
